CLDR-14668 Fix Coverage bugs in modern front end; preliminary refactoring
-New cldrCoverage.js with code moved from cldrSurvey.js

-New subroutine getMenusAjaxUrl in cldrMenus.js

-Comments
